+++ spec/channel/branding_settings_spec.rb
@@ -10,6 +10,8 @@ describe 'Yt::Channel’s branding settings methods', :server do
       expect(channel.banner_image_url).to be_a String
       expect(channel.keywords).to match_array ['YouTube', 'channel', 'test']
       expect(channel.unsubscribed_trailer).to eq 'gknzFj_0vvY'
+      expect(channel.featured_channels_title).to eq 'Public featured channels'
+      expect(channel.featured_channels_urls).to eq %w(UCxO1tY8h1AhOz0T4ENwmpow)
     end
   end